implement hyperion restart. core function is good, but needs a beeter structuring- something for next refactoring session ;-)
