+++ assets/webconfig/index.html
@@ -225,10 +225,10 @@
 
 		<!-- Page Content -->
 		<div id="page-wrapper" style="padding-top:10px">
-			<div id="hyperion_restart_notify" class="alert alert-warning" style="display:none;padding:10px;margin:0">
+			<div id="hyperion_reload_notify" class="alert alert-warning" style="display:none;padding:10px;margin:0">
 			<div class="panel-danger" style="text-align:right">
 				<div style="float:left">Hyperion Configuration is modified. To make it active, restart Hyperion.</div>
-				<button id="btn_hyperion_restart" class="btn btn-danger">Restart Hyperion</button>
+				<button id="btn_hyperion_reload" class="btn btn-danger">Restart Hyperion</button>
 			</div>
 		</div>
