+++ assets/webconfig/js/content_index.js
@@ -22,9 +22,9 @@ $(document).ready( function() {
 		cleanCurrentVersion = currentVersion.replace(/\./g, '');
 
 		if (parsedServerInfoJSON.info.hyperion[0].config_modified)
-			$("#hyperion_restart_notify").fadeIn("fast");
+			$("#hyperion_reload_notify").fadeIn("fast");
 		else
-			$("#hyperion_restart_notify").fadeOut("fast");
+			$("#hyperion_reload_notify").fadeOut("fast");
 
 		// get active led device
 		var leddevice = parsedServerInfoJSON.info.ledDevices.active;
@@ -75,6 +75,13 @@ $(document).ready( function() {
 		requestServerInfo();
 	});
 
+	$("#btn_hyperion_reload").on("click", function(){
+		$(hyperion).off();
+		requestServerConfigReload();
+		watchdog = 1;
+		$("#wrapper").fadeOut("slow");
+		cron();
+	});
 });
 
 $(function(){
